Finish visualizer library conversion
Fix all warnings. Re-implement single animated SVGs. Fix and/or rewrite
unit tests, integration tests and benchmarks. Clean up comments and add
some documentation for the public functions.

Did not re-implement the multi-svg animations; I do not know that there
is a good use case for this. Nonetheless, the code remains commented
out, although maybe better to delete if we aren't going to use it.

The main.rs file is now more of a demo than anything, but still works
with command line arguments:
```
Usage: holoviz [OPTIONS] <INPUT_SVG> <OUTPUT_SVG> [DURATION]

Arguments:
  <INPUT_SVG>   Input file: expects a .svg with circles
  <OUTPUT_SVG>  Output file: expects .svg
  [DURATION]    Animation duration in seconds [default: 2]

Options:
  -a, --animate  Animate the output
  -h, --help     Print help
  -V, --version  Print version
```
